+++ doc/Changelog.md
@@ -1,4 +1,4 @@
-* albums can be protected with passwords, which aren't exposed in the album tree or the cache
+* albums can be protected with passwords, which aren't exposed neither in the album tree nor the cache
 * new options for passwords management: `passwords_marker`, `passwords_file`
 * when showing a map album, `esc` takes back to the map and the popup
 * right menu: bug fixes and better ordering of the options